pyproject.toml
[project]
name = "text-based-adventure-game"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
]
